+++ .gitignore
@@ -4,3 +4,9 @@
 deps/deps.jl
 
 Manifest.toml
+docs/Manifest.toml
+docs/build
+docs/src/generated
+docs/src/misc/contributing.md
+benchmark/tune.json
+benchmark/results*
