+++ docs/src/index.md
@@ -1,0 +1,4 @@
+# ManifoldDiffEq
+
+The package __ManifoldDiffEq__ aims to provide a library of differential equation solvers
+on manifolds. The library is built on top of [`Manifolds.jl`](https://github.com/JuliaManifolds/Manifolds.jl) and follows the interface of [`OrdinaryDiffEq.jl`](https://github.com/SciML/OrdinaryDiffEq.jl/).
